feat: harden inventory ledger workflows

+++ app/Exceptions/InsufficientInventoryException.php
@@ -6,9 +6,9 @@ use Exception;
 
 class InsufficientInventoryException extends Exception
 {
-    public int $available;
+    public float $available;
 
-    public function __construct(int $available)
+    public function __construct(float $available)
     {
         parent::__construct('Insufficient inventory available.');
         $this->available = $available;
